requirements.txt
pandas==3.0.1
pykan==0.2.8
scikit-learn==1.8.0
torch==2.10.0+cpu
tqdm==4.67.3
matplotlib==3.10.8
PyYAML==6.0.3